+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 * An --insecure (INSECURE=yes) option is added that mimics curl
   and allows bypassing SSL certifiation verification
 
+* Fix hooks (e.g. before_each) not working properly when using
+  configuration folders.
+
 # 0.18.0 - 2021/02/23
 
 * Dependencies upgraded, http >= 4.0 and finitio >= 0.10.0